change on todoContext
